pyproject.toml
[tool.poetry]
name = "coursework-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.6"
psycopg2 = "^2.9.10"
pillow = "^11.3.0"
djangorestframework = "^3.16.1"
dotenv = "^0.9.9"
djangorestframework-simplejwt = "^5.5.1"
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
